Updated install_labels function
Similar to changes on dropping, updated install labels etc. functions to work with the new format as well. Also changed functions to accept an identical, existing constraint or index as successfully created; this primarily solved an issue with tests trying to recreate the same constraints (due to how tests executes), but also in the event that a user runs the install_label command multiple times.