Updated multiprocessing tests
Updated multiprocessing tests, to use a threadpool instead of a process pool. Pytests does not play nice with multiprocessing, when it comes to retaining connections. Outside of pytest, multiprocessing works fine (both from a single script, and via Celery).
Also ensured that the pool of the tests are properly closed after run.