Add support for slow query log (#480)
* Updating utils.py to support a slow query timer

Adds support so that logging only occurs if the query time (``tte``) exceeds a threshold set by a new environmental setting.

Example usage: Setting ```os.environ["NEOMODEL_SLOW_QUERIES"] = "0.09"``` will cause only queries that take longer than 0.09 seconds to run to be logged. If NEOMODEL_SLOW_QUERIES is not set, then it is treated as being 0 and therefore the logging acts as before

* Updating utils.py to support a slow query timer

Adds support so that logging only occurs if the query time (``tte``) exceeds a threshold set by a new environmental setting.

Example usage: Setting ```os.environ["NEOMODEL_SLOW_QUERIES"] = "0.09"``` will cause only queries that take longer than 0.09 seconds to run to be logged. If NEOMODEL_SLOW_QUERIES is not set, then it is treated as being 0 and therefore the logging acts as before