Search frames for class in RelationDefinition
Currently, the RelationshipDefinition uses a hard coded frame number to
retrieve the node class used for a relationship. Depending on the stack
layout, this might be the wrong frame. Therefore use Python's inspect
module to search all outer frames for the node class and use the number
of the the frame with the first match.